+++ R/cc_cen.R
@@ -64,7 +64,7 @@ cc_cen <- function(x,
   }
 
   # set default projection
-  wgs84 <- "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
+  wgs84 <- "+proj=longlat +datum=WGS84 +no_defs"
   
   # select relevant columns
   dat <- sp::SpatialPoints(x[, c(lon, lat)],