+++ R/cc_urb.R
@@ -76,7 +76,7 @@ cc_urb <- function(x,
   }
 
   # Prepare input points and extent
-  wgs84 <- "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
+  wgs84 <- "+proj=longlat +datum=WGS84 +no_defs"
 
   dat <- sp::SpatialPoints(x[, c(lon, lat)], proj4string = CRS(wgs84))
   limits <- raster::extent(dat) + 1